fix: pr requests and lint problems
